+++ src/Phan/Analysis/ParameterTypesAnalyzer.php
@@ -11,7 +11,6 @@ use Phan\Language\Element\FunctionInterface;
 use Phan\Language\Element\Method;
 use Phan\Language\Element\Parameter;
 use Phan\Language\FQSEN\FullyQualifiedClassName;
-use Phan\Language\Type\GenericArrayType;
 use Phan\Language\Type\IterableType;
 use Phan\Language\Type\MixedType;
 use Phan\Language\Type\NullType;
@@ -48,17 +47,9 @@ class ParameterTypesAnalyzer
             $union_type = $parameter->getUnionType();
 
             // Look at each type in the parameter's Union Type
-            foreach ($union_type->withFlattenedArrayShapeTypeInstances()->getTypeSet() as $outer_type) {
-                $type = $outer_type;
-
-                // TODO: Add unit test of `array{key:MissingClazz}`
-                while ($type instanceof GenericArrayType) {
-                    $type = $type->genericArrayElementType();
-                }
-
-                // If its a native type or a reference to
-                // self, its OK
-                if ($type->isNativeType() || ($method instanceof Method && ($type->isSelfType() || $type->isStaticType()))) {
+            foreach ($union_type->getReferencedClasses() as $outer_type => $type) {
+                // If it's a reference to self, its OK
+                if ($method instanceof Method && ($type->isSelfType() || $type->isStaticType())) {
                     continue;
                 }